Change nbsteps() to uint32_t to make it independent from 32/64-bit platform.

+++ src/control.cc
@@ -78,7 +78,7 @@ void ControlModule::for_eval(AbstractNode &node, const ModuleInstantiation &inst
 		Context c(ctx);
 		if (it_values.type() == Value::RANGE) {
 			Value::RangeType range = it_values.toRange();
-                        unsigned long steps = range.nbsteps();
+                        uint32_t steps = range.nbsteps();
                         if (steps >= 10000) {
                                 PRINTB("WARNING: Bad range parameter in for statement: too many elements (%lu).", steps);
                         } else {
@@ -229,7 +229,7 @@ AbstractNode *ControlModule::instantiate(const Context* /*ctx*/, const ModuleIns
 			else if (value.type() == Value::RANGE) {
 				AbstractNode* node = new AbstractNode(inst);
 				Value::RangeType range = value.toRange();
-                                unsigned long steps = range.nbsteps();
+                                uint32_t steps = range.nbsteps();
 				if (steps >= 10000) {
 					PRINTB("WARNING: Bad range parameter for children: too many elements (%lu).", steps);
 					return NULL;
